Fixed circular dependencies between metadata_query.ts to filters/utils.ts in Salesforce Adapter.
