+++ packages/salesforce-adapter/test/fetch_profile/data_management.test.ts
@@ -14,13 +14,14 @@
 * limitations under the License.
 */
 
-import { buildDataManagement, DataManagement } from '../../src/fetch_profile/data_management'
+import { buildDataManagement } from '../../src/fetch_profile/data_management'
 import {
   API_NAME,
   DETECTS_PARENTS_INDICATOR,
 } from '../../src/constants'
 import { createCustomObjectType } from '../utils'
 import { Types } from '../../src/transformers/transformer'
+import { DataManagement } from '../../src/types'
 
 describe('buildDataManagement', () => {
   let dataManagement: DataManagement
